+++ .gitignore
@@ -1,3 +1,6 @@
+*.db
+venv/
+
 # Created by .ignore support plugin (hsz.mobi)
 ### Python template
 # Byte-compiled / optimized / DLL files
